refactor(service tiles): reduce duplication
- crated new `ServiceTile` that combine duplicated code
  for `Service` and `ServiceTiles` with condition for coloring
  based on `hasDiagonalGradientBackground`

SVA-639
